requirements.txt
tensorflow-gpu
tf-nightly
keras
Pillow
numpy
opencv-python
uuid
flask
opencv
matplotlib